+++ src/vidore_benchmark/retrievers/colqwen2_retriever.py
@@ -32,6 +32,7 @@ class ColQwen2Retriever(BaseVisionRetriever):
         pretrained_model_name_or_path: str = "vidore/colqwen2-v1.0",
         device: str = "auto",
         num_workers: int = 0,
+        **kwargs,
     ):
         super().__init__(use_visual_embedding=True)